requirements.txt
django
psycopg2
Pillow
crispy-bootstrap5
django-crispy-forms
djangorestframework
environs